+++ match-service/src/main/java/com/lgcns/bebee/match/presentation/dto/req/AgreementCreateReqDTO.java
@@ -22,8 +22,8 @@ public class AgreementCreateReqDTO {
     private EngagementType type;
     private Boolean isVolunteer;
     private List<Long> helpCategoryIds;
-    private Integer unitHoney;
-    private Integer totalHoney;
+    private Long unitHoney;
+    private Long totalHoney;
     private String region;
     private String chatroomId;
